pull out extract functions for integrated corpus
