+++ scripts/cxg_admin_scripts/delete.py
